restructure denoiser C code, so that the core functions of denoiser are in new files.
