+++ mbircone/src/denoise3D.h
@@ -1,0 +1,5 @@
+#include "MBIRModularUtilities3D.h"
+#include "icd3d.h"
+#include "icd3dDenoise.h"
+
+void MBIR3DDenoise(struct Image *img, struct Image *img_noisy, struct ReconParams *reconParams);
