+++ mbircone/src/recon3DCone.h
@@ -3,5 +3,3 @@
 
 
 void MBIR3DCone(struct Image *img, struct Sino *sino, struct ReconParams *reconParams, struct SysMatrix *A);
-
-void MBIR3DDenoise(struct Image *img, struct Image *img_noisy, struct ReconParams *reconParams);
