+++ setup.py
@@ -17,6 +17,7 @@ PACKAGE_DIR = "mbircone"
 
 SRC_FILES = [PACKAGE_DIR + '/src/allocate.c', PACKAGE_DIR + '/src/MBIRModularUtilities3D.c',
              PACKAGE_DIR + '/src/icd3d.c', PACKAGE_DIR + '/src/recon3DCone.c',
+             PACKAGE_DIR + '/src/icd3dDenoise.c', PACKAGE_DIR + '/src/denoise3D.c',
              PACKAGE_DIR + '/src/computeSysMatrix.c',
              PACKAGE_DIR + '/src/interface.c', PACKAGE_DIR + '/interface_cy_c.pyx']
